Clarify how to sign in using the Service Principal
It is confusing the way it is explained how to sing in into Azure using the Service Principal. It is supposed that this logins is going to be made in another Power Shell session so variables such as $svcprincipal.ApplicationId won't be available.

Another point is that sometimes using the service principal for login is made to avoid interaction with a user and make an automation task; I am suggesting to add an example of how this could be done.